CornHarvester Update and a few Usability Changes
CornHarvester is the Mass-Domain Usage version of TheHarvester Tool in Kali Linux. It will search a wordlist of domains to continually run email addresses searches on for Spear-Phishing Attacks

Added a new How-To-Video

Added a few design changes for Reconnaissance, in Option #1 for AC. Now it will automatically save a recon report in your /root/ArmsCommander/logs directory. I did this to make sure that it stops nagging you for "too many questions"