Pupy Integration Test Release #1
Double Checking. Pupy has a really funny way of installing. I really would love to integrate it for you guys but, I just learned that the installer generates sensitive crypto keys that I cannot risk losing to anyone. That is a part of the Pupy installation process apparently.

Still in testing.