The Second Prune (Not Required Update)
For usability purposes i performed a second prune to reorganize the Remote Exploits Menu. It was starting to get disorganized and ugly.

There are only three categories now

>1. Payload Generators/Listeners

>2. Toolkits

>3. Other Remote Exploitation Stuff

There is NO change in usability. But from now on, this is what AC will look like for future integration with other software. Easier on your eyes.